Fix styles, rename hook funtction
